pyproject.toml
[project]
name = "my-notion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "notion-client>=2.0.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.6",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.0",
    "ruff>=0.14.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
